split itvx and itv legacy code into separate modules
