+++ plugin.video.itvhub/resources/lib/itv.py
@@ -59,55 +59,6 @@ def get_live_schedule(hours=4):
     return schedule
 
 
-def get_live_channels():
-    from tzlocal import get_localzone
-    local_tz = get_localzone()
-    utc_tz = pytz.utc
-
-    live_data = fetch.get_json('https://nownext.oasvc.itv.com/channels?broadcaster=itv&featureSet=mpeg-dash,clearkey,outband-webvtt,hls,aes,playready,widevine,fairplay&platformTag=dotcom')
-    fanart_url = live_data['images']['backdrop']
-
-    main_schedule = get_live_schedule()
-
-    channels = live_data['channels']
-
-    for channel in channels:
-        channel['backdrop'] = fanart_url
-        slots = channel.pop('slots')
-
-        # The itv main live channels get their schedule from the full live schedule
-        if channel['channelType'] == 'simulcast':
-            chan_id = channel['id']
-            for main_chan in main_schedule:
-                # Caution, might get broken when ITV becomes ITV1 everywhere
-                if main_chan['channel']['name'] == chan_id:
-                    channel['slot'] = main_chan['slot']
-                    break
-            if channel.get('slot'):
-                # On to the next channel if adding full schedule succeeded
-                continue
-
-        programs_list = []
-        for prog in (slots['now'], slots['next']):
-            if prog['detailedDisplayTitle']:
-                details = ': '.join((prog['displayTitle'], prog['detailedDisplayTitle']))
-            else:
-                details = prog['displayTitle']
-
-            start_t = prog['start'][:19]
-            # TODO: check this in DST period
-            utc_start = datetime(*(time.strptime(start_t, '%Y-%m-%dT%H:%M:%S')[0:6])).replace(tzinfo=utc_tz)
-
-            programs_list.append({
-                'programme_details': details,
-                'programmeTitle': prog['displayTitle'],
-                'orig_start': None,          # fast channels do not support play from start
-                'startTime': utc_start.astimezone(local_tz).strftime('%H:%M')
-            })
-        channel['slot'] = programs_list
-    return channels
-
-
 stream_req_data = {
     'client': {
         'id': 'browser',
@@ -407,28 +358,6 @@ def productions(url, show_name):
     return series_list
 
 
-def get_playlist_url_from_episode_page(page_url):
-    """Obtain the url to the episode's playlist from the episode's HTML page.
-    """
-    import re
-
-    logger.info("Get playlist from episode page - url=%s", page_url)
-    html_doc = fetch.get_document(page_url)
-    logger.debug("successfully retrieved page %s", page_url)
-
-    # New version - might a bit overdone as just a regex to obtain the playlist url should suffice.
-    # doc_data = parse.get__next__data_from_page(html_doc)
-    # player_data = doc_data['props']['pageProps']['episodeHeroWrapperProps']['playerProps']
-    # name = player_data['programmeTitle']
-    # play_list_url = player_data['playlistUrl']
-
-    # Only this will fail on itvX, but is the name actually used anywhere?
-    # name = re.compile('data-video-title="(.+?)"').search(html_doc)[1]
-    name = ''
-    play_list_url = re.compile('data-video-id="(.+?)"').search(html_doc)[1]
-    return play_list_url, name
-
-
 def get_vtt_subtitles(subtitles_url):
     show_subtitles = Script.setting['subtitles_show'] == 'true'
     if show_subtitles is False:
@@ -457,66 +386,3 @@ def get_vtt_subtitles(subtitles_url):
         logger.error("Failed to get vtt subtitles from url %s", subtitles_url, exc_info=True)
         return None
 
-
-def search(search_term):
-    url = 'https://textsearch.prd.oasvc.itv.com/search'
-    query_params = {
-        'broadcaster': 'itv',
-        'featureSet': 'clearkey,outband-webvtt,hls,aes,playready,widevine,fairplay,bbts,progressive,hd,rtmpe',
-        # We can handle only free items because of the way we list production right now.
-        'onlyFree': 'true',
-        'platform': 'dotcom',
-        'query': search_term
-    }
-    data = fetch.get_json(url, params=query_params)
-    if data is None:
-        return
-
-    results = data.get('results')
-
-    def parse_programme(prg_data):
-        prog_name = prg_data['programmeTitle']
-        img_url = prg_data['latestAvailableEpisode']['imageHref']
-
-        return {
-            'entity_type': 'programme',
-            'label': prog_name,
-            'art': {'thumb': img_url.format(width=960, height=540, quality=80, blur=0, bg='false')},
-            'info': {'plot': prg_data.get('synopsis'),
-                     'title': '[B]{}[/B] - {} episodes'.format(prog_name, prg_data.get('totalAvailableEpisodes', ''))},
-            'params': {
-                'url': 'https://discovery.hubsvc.itv.com/platform/itvonline/dotcom/productions?programmeId={}&'
-                       'features=aes,clearkey,fairplay,hls,mpeg-dash,outband-webvtt,playready,'
-                       'widevine&broadcaster=itv'.format(prg_data['legacyId']['apiEncoded']),
-                'name': prog_name}
-        }
-
-    def parse_special(prg_data):
-        # AFAICT special is always a production, which might be a production of a programme, but
-        # presented as a single episode in the search results.
-        prog_name = program_data['specialTitle']
-        img_url = program_data['imageHref']
-        # convert productionId to a format used in the url
-        api_prod_id = prg_data['productionId'].replace('/', '_').replace('#', '.')
-
-        return {
-            'entity_type': 'special',
-            'label': prog_name,
-            'art': {'thumb': img_url.format(width=960, height=540, quality=80, blur=0, bg='false')},
-            'info': {'plot': prg_data.get('synopsis'),
-                     'title': prog_name},
-            'params': {'url': 'https://magni.itv.com/playlist/itvonline/ITV/' + api_prod_id,
-                       'name': prog_name}
-        }
-
-    for result in results:
-        program_data = result['data']
-        entity_type = result['entityType']
-
-        if entity_type == 'programme':
-            yield parse_programme(program_data)
-        elif entity_type == 'special':
-            yield parse_special(program_data)
-        else:
-            logger.warning("Unknown search result item entityType %s on search term %s", entity_type, search_term)
-            continue
