+++ test/local/test_itv.py
@@ -32,16 +32,3 @@ from resources.lib import itv
 setUpModule = fixtures.setup_local_tests
 tearDownModule = fixtures.tear_down_local_tests
 
-
-class Search(TestCase):
-    @patch('resources.lib.fetch.get_json', return_value=open_json('search/the_chase.json'))
-    def test_simple_search(self, _):
-        result = itv.search('the_chase')
-        self.assertIsInstance(result, types.GeneratorType)
-        self.assertEqual(10, len(list(result)))
-
-    @patch('resources.lib.fetch.get_json', return_value=None)
-    def test_search_without_results(self, _):
-        result = itv.search('xprs')
-        self.assertIsInstance(result, types.GeneratorType)
-        self.assertListEqual(list(result), [])
